Fix Clang 16 compiler warning: format (1) (#335)
While compiling Clang 16 generates the following warnings.

in spine.c:

```
spine.c:934:101: warning: format specifies type 'long' but the argument has type 'pthread_t' (aka 'struct pthread *') [-Wformat]
                                SPINE_LOG_DEBUG(("DEBUG: Device[%i] Valid Thread to be Created (%ld)", poller_details->host_id, threads[device_counter]));
```

in util.c:
```
util.c:1265:100: warning: format specifies type 'long' but the argument has type 'pthread_t' (aka 'struct pthread *') [-Wformat]
        snprintf(logprefix, SMALL_BUFSIZE, "SPINE: Poller[%i] PID[%i] PT[%ld] ", set.poller_id, getpid(), pthread_self());
```
This one i'm not as sure about, maybe this should be cast to a long
unsigned int instead? I have not had problems running it this way but
i'm not 100% sure this the right fix.

+++ util.c
@@ -1263,7 +1263,7 @@ int spine_log(const char *format, ...) {
 	fp = stdout;
 
 	/* log message prefix */
-	snprintf(logprefix, SMALL_BUFSIZE, "SPINE: Poller[%i] PID[%i] PT[%ld] ", set.poller_id, getpid(), pthread_self());
+	snprintf(logprefix, SMALL_BUFSIZE, "SPINE: Poller[%i] PID[%i] PT[%ld] ", set.poller_id, getpid(), (long int)pthread_self());
 
 	/* get time for poller_output table */
 	nowbin = time(&nowbin);
